rename set extensions, fix geometry constructions, remove unnecessary "mark as used" hacks, add another Nes constructor, add flatTraverse syntax, refactor Guid construction, sbt 1.3.10, update dependencies

+++ modules/base/src/main/scala/scutil/geom/IntPoint.scala
@@ -47,7 +47,8 @@ final case class IntPoint(x:Int, y:Int) {
 
 	def angle:Double	= smath atan2 (y, x)
 
-	def length:Double	= smath sqrt lengthQ
+	// TODO this is dangerous, converting Long to Double is lossy
+	def length:Double	= smath sqrt lengthQ.toDouble
 	def lengthQ:Long	= x*x + y*y
 
 	//------------------------------------------------------------------------------
